# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RAGO API
  description: |
    RAGO is a local-first, privacy-first Go AI foundation with four core pillars:
    - **LLM**: Language model management and generation
    - **RAG**: Retrieval-augmented generation
    - **MCP**: Model Context Protocol for tool integration
    - **Agents**: Autonomous workflow orchestration
  version: 3.0.0
  contact:
    name: RAGO Support
    url: https://github.com/liliang-cn/rago
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:7127/api/v1
    description: Local development server
  - url: https://api.rago.ai/v1
    description: Production server
...
paths:
...
  /agents/workflows:
    get:
      tags: [Agents]
      summary: List workflows
      description: Get a list of all defined workflows
      security:
        - bearerAuth: []
        - apiKey: []
      responses:
        '200':
          description: Workflows listed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Agents]
      summary: Create workflow
      description: Create a new workflow definition
      security:
        - bearerAuth: []
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowDefinition'
      responses:
        '201':
          description: Workflow created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /agents/workflows/{name}/execute:
    post:
      tags: [Agents]
      summary: Execute workflow
      description: Execute a workflow with the provided parameters
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Workflow name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowRequest'
      responses:
        '200':
          description: Workflow executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # Unified endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    WorkflowInfo:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        steps:
          type: integer
    WorkflowDefinition:
      type: object
      required: [name, steps]
      properties:
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            type: object
    WorkflowRequest:
      type: object
      properties:
        parameters:
          type: object
    WorkflowResponse:
      type: object
      properties:
        result:
          type: object
        status:
          type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string